JD: Commiting scraping along with front end
Requires changes to regex in the Room class so that it fixes the time as military time (add 12 to the time if pm). Create the JSON writing method reading method. Needs a Time/Date class to hold the times by days and then times; this will need the Room class to be changed to allow for this change.